+++ udata_search_service/__init__.py
@@ -7,4 +7,4 @@ import os
 tag = os.environ.get('CIRCLE_TAG')
 build_num = os.environ.get('CIRCLE_BUILD_NUM')
 
-__version__ = '2.1.1.dev' + (str(build_num) if not tag and build_num else '')
+__version__ = '2.2.0' + (str(build_num) if not tag and build_num else '')